+++ changelog.md
@@ -1,5 +1,11 @@
 # Tic Tac Toe Changelog
 
+## [1.1.0] - 2021-09-20
+### Added
+- Choose opponent modal
+### Edited
+- Restart logic
+
 ## [1.0.0] - 2021-09-20
 ### Added
 - Usable tic tac toe